# backend/docs/swagger.json
{
    "openapi": "3.0.2",
    "info": {
        "version": "1.0",
        "title": "OnnelikudAutod",
        "description": "### This is a Swagger UI for our OnnelikudAutod API  \n\n### Error handling\nErrors are presented using the HTTP status codes 400-599. The response includes a more detailed explanation about the error in the body.\n\n```\n{\n  \"error\": \"Cleartext explanation of the error that occurred\"\n}\n```\n\n### Generic API errors\nThese error responses can be returned by any of the described endpoints.\n\n| HTTP status  |  Description |\n|---|---|\n| 400| Bad input parameters or invalid request. More information can be found in the response body.\n| 404| The requested resource can not be found."
    },
    "paths": {
        "/onnelikud-autod": {
            "summary": "OnnelikudAutod Object",
            "get": {
                "tags": [
                    "OnnelikudAutod"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                },
                                "examples": {
                                    "Cars": {
                                        "value": [
                                            {"id": 1, "name": "Toyota Corolla"},
                                            {"id": 2, "name": "Honda Civic"},
                                            {"id": 3, "name": "Ford Focus"}
                                        ]
                                    },
                                    "None": {
                                        "value": []
                                    }
                                }
                            }
                        },
                        "description": "Korras"
                    }
                },
                "summary": "Kõik autod"
            },
            "post": {
                "tags": [
                    "OnnelikudAutod"
                ],
                "summary": "Lisa uus auto",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "description": "Auto nimi mida lisad"
                                    },
                                    "price": {
                                        "type": "number",
                                        "description": "Auto hind"
                                    }
                                },
                                "required": ["name"]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Auto lisatud edukalt",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Auto lisatud edukalt"
                                        },
                                        "car": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "price": {
                                                    "type": "number"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad input parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Missing required field 'name'."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/onnelikud-autod/{id}": {
            "get": {
                "tags": [
                    "OnnelikudAutod"
                ],
                "summary": "Leia auto ID järgi",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Auto leitud",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "price": {
                                            "type": "number"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ebakorralik ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "ID peab olema ainult numbrid"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Autot ei leitud",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Autot ei leitud"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}